Take more projection into account
